+++ api/src/api/routes/contracts.py
@@ -1,4 +1,5 @@
 from fastapi import APIRouter
+from fastapi import HTTPException
 from fastapi import Query
 from fastapi import Request
 
@@ -43,5 +44,35 @@ async def get_contract_address_count(
         query += f" and value < ${len(args)}"
 
     async with request.app.state.db.acquire() as conn:
-        res = await conn.fetchrow(query, *args)
-        return res["cnt"]
+        row = await conn.fetchrow(query, *args)
+    if row["cnt"] is None:
+        raise HTTPException(status_code=404)
+    return row["cnt"]
+
+
+@r.get("/supply", description="Supply in contracts")
+async def supply_in_contracts(
+    request: Request,
+    token_id: TokenID = Query(None, description="Optional token id"),
+):
+    """
+    Current supply in contract addresses. Excludes coinbase address.
+    """
+    query = """
+        select sum(value) as value
+        from bal.erg
+        where address <> '2Z4YBkDsDvQj8BX7xiySFewjitqp2ge9c99jfes2whbtKitZTxdBYqbrVZUvZvKv6aqn9by4kp3LE1c26LCyosFnVnm6b6U1JYvWpYmL2ZnixJbXLjWAWuBThV1D6dLpqZJYQHYDznJCk49g5TUiS4q8khpag2aNmHwREV7JSsypHdHLgJT7MGaw51aJfNubyzSKxZ4AJXFS27EfXwyCLzW1K6GVqwkJtCoPvrcLqmqwacAWJPkmh78nke9H4oT88XmSbRt2n9aWZjosiZCafZ4osUDxmZcc5QVEeTWn8drSraY3eFKe8Mu9MSCcVU' 
+            and address <> '4L1ktFSzm3SH1UioDuUf5hyaraHird4D2dEACwQ1qHGjSKtA6KaNvSzRCZXZGf9jkfNAEC1SrYaZmCuvb2BKiXk5zW9xuvrXFT7FdNe2KqbymiZvo5UQLAm5jQY8ZBRhTZ4AFtZa1UF5nd4aofwPiL7YkJuyiL5hDHMZL1ZnyL746tHmRYMjAhCgE7d698dRhkdSeVy'
+            and (address not like '9%' or length(address) <> 51)
+    """
+    args = []
+    if token_id is not None:
+        args.append(token_id)
+        query = query.replace("bal.erg", "bal.tokens")
+        query += f" and token_id = $1"
+
+    async with request.app.state.db.acquire() as conn:
+        row = await conn.fetchrow(query, *args)
+        if row["value"] is None:
+            raise HTTPException(status_code=404)
+        return row["value"]