+++ api/src/tests/test_tokens.py
@@ -6,12 +6,13 @@ from ..main import app
 from .db import MockDB
 
 TOKEN_A = "tokenaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-TOKEN_B = "tokenbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
+TOKEN_B = "tokenbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbeip4"
 TOKEN_X = "validxtokenxidxofxnonxexistingxtokenxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
 
 
 @pytest.fixture(scope="module")
 def client():
+    P2PK = "9" * 51
     sql = f"""
         insert into core.headers (height, id, parent_id, timestamp) values 
         (10, 'header10', 'header09', 1567123456789),
@@ -27,8 +28,9 @@ def client():
         ('box-2', 'tx-2', 'header20', 20, 'addr1', 0, 10000000);
         
         insert into core.tokens (id, box_id, emission_amount) values
-        ('{TOKEN_A}', 'box-1', 900),
-        ('{TOKEN_B}', 'box-2', 800);
+        ('{TOKEN_A}', 'box-1', 900);
+        insert into core.tokens (id, box_id, emission_amount, name, description, decimals, standard) values
+        ('{TOKEN_B}', 'box-2', 800, 'token_b', 'description of token b', 2, 'EIP-4');
 
         insert into bal.tokens_diffs (address, token_id, height, tx_id, value) values
         ('addr1', '{TOKEN_A}', 10, 'tx_1',   900),
@@ -36,17 +38,50 @@ def client():
         ('addr1', '{TOKEN_A}', 20, 'tx_2',  -200),
         ('addr2', '{TOKEN_A}', 20, 'tx_2',   150),
         ('addr1', '{TOKEN_A}', 30, 'tx_3',  -300),
-        ('addr3', '{TOKEN_A}', 30, 'tx_3',   300);
+        ('{P2PK}', '{TOKEN_A}', 30, 'tx_3',   300);
 
         insert into bal.tokens (address, token_id, value) values
         ('addr1', '{TOKEN_A}', 400),
         ('addr1', '{TOKEN_B}', 800),
         ('addr2', '{TOKEN_A}', 150),
-        ('addr3', '{TOKEN_A}', 300);
+        ('{P2PK}', '{TOKEN_A}', 300);
     """
     with MockDB(sql=sql) as _:
         with TestClient(app) as client:
             yield client
+
+
+class TestDetails:
+    def test_dummy_token(self, client):
+        url = f"/tokens/{TOKEN_A}"
+        response = client.get(url)
+        assert response.status_code == 200
+        assert response.json() == {
+            "token_id": TOKEN_A,
+            "emission_amount": 900,
+            "name": None,
+            "description": None,
+            "decimals": 0,
+            "standard": None,
+        }
+
+    def test_eip4_token(self, client):
+        url = f"/tokens/{TOKEN_B}"
+        response = client.get(url)
+        assert response.status_code == 200
+        assert response.json() == {
+            "token_id": TOKEN_B,
+            "name": "token_b",
+            "description": "description of token b",
+            "emission_amount": 800,
+            "decimals": 2,
+            "standard": "EIP-4",
+        }
+
+    def test_unknown_token(self, client):
+        url = f"/tokens/{TOKEN_X}"
+        response = client.get(url)
+        assert response.status_code == 404
 
 
 class TestSupply:
@@ -55,8 +90,9 @@ class TestSupply:
         response = client.get(url)
         assert response.status_code == 200
         assert response.json() == {
-            "total": 900,
-            "circulating": 850,
+            "emitted": 900,
+            "in_p2pks": 300,
+            "in_contracts": 550,
             "burned": 50,
         }