+++ watcher/src/main.rs
@@ -11,13 +11,6 @@ use session::Session;
 fn main() -> Result<(), &'static str> {
     let mut session = Session::new()?;
 
-    if session.db.is_empty().unwrap() {
-        if !session.allow_bootstrap {
-            session.db.apply_constraints_all().unwrap();
-        }
-        sync::include_genesis_boxes(&session)?;
-    };
-
     if session.allow_bootstrap {
         let db_constraints_status = session.db.constraints_status().unwrap();
         if !db_constraints_status.tier_1 {