+++ watcher/testbench/test_migrations.py
@@ -1,10 +1,12 @@
+import re
 import pytest
 import psycopg as pg
 
 from fixtures.api import MockApi, ApiUtil, GENESIS_ID
 from fixtures.config import temp_cfg
-from fixtures.db import bootstrap_db
-from fixtures.db import temp_db
+from fixtures.db import temp_db_rev1
+from fixtures.db import SCHEMA_PATH
+from fixtures.db import fill_rev1_db
 from fixtures.addresses import AddressCatalogue as AC
 from utils import run_watcher
 
@@ -16,12 +18,6 @@ def make_blocks(height: int):
     block a - coinbase tx:
         base-box1 1000 --> base-box2  950
                            con1-box1   50
-
-    ----------------------fork-of-b----------------------
-    block x - fork of block b to be ignored/rolled back:
-        con1-box1   50 --> con9-box1   30
-                           pub9-box1   20 (3000 con1-box1)
-    ------------------------------------------------------
 
     block b - minting a token and using registers:
         con1-box1   50 --> con2-box1   40
@@ -76,44 +72,6 @@ def make_blocks(height: int):
     }
 
     h += 1
-    tx_x1 = {
-        "id": "tx-x1",
-        "inputs": [{"boxId": "con1-box1"}],
-        "dataInputs": [],
-        "outputs": [
-            {
-                "boxId": "con9-box1",
-                "value": 30,
-                "ergoTree": con2.ergo_tree,
-                "assets": [],
-                "creationHeight": h,
-                "additionalRegisters": {
-                    "R4": "0703553448c194fdd843c87d080f5e8ed983f5bb2807b13b45a9683bba8c7bfb5ae8",
-                    "R5": "0e2098479c7d306cccbd653301102762d79515fa04c6f6b35056aaf2bd77a7299bb8",
-                    "R6": "05a4c3edd9998877",
-                },
-                "transactionId": "tx-x1",
-                "index": 0,
-            },
-            {
-                "boxId": "pub9-box1",
-                "value": 20,
-                "ergoTree": pub1.ergo_tree,
-                "assets": [
-                    {
-                        "tokenId": "con1-box1",
-                        "amount": 3000,
-                    }
-                ],
-                "creationHeight": h,
-                "additionalRegisters": {},
-                "transactionId": "tx-x1",
-                "index": 1,
-            },
-        ],
-        "size": 674,
-    }
-
     tx_b1 = {
         "id": "tx-b1",
         "inputs": [{"boxId": "con1-box1"}],
@@ -243,24 +201,6 @@ def make_blocks(height: int):
         "size": 1000,
     }
 
-    block_x = {
-        "header": {
-            "votes": "000000",
-            "timestamp": 1234560200000,
-            "size": 123,
-            "height": height + 2,
-            "id": "block-x",
-            "parentId": "block-a",
-        },
-        "blockTransactions": {
-            "headerId": "block-x",
-            "transactions": [tx_x1],
-            "blockVersion": 2,
-            "size": 1155,
-        },
-        "size": 1000,
-    }
-
     block_b = {
         "header": {
             "votes": "000000",
@@ -297,7 +237,7 @@ def make_blocks(height: int):
         "size": 1000,
     }
 
-    return [block_a, block_x, block_b, block_c]
+    return [block_a, block_b, block_c]
 
 
 @pytest.mark.order(5)
@@ -316,15 +256,16 @@ class TestMigrations:
             yield api
 
     @pytest.fixture()
-    def db(self, api, temp_db):
+    def db(self, api, temp_db_rev1):
         """
         Loads mock api and provides test db connection.
 
         Function scoped fixture since tests will be changing the db state.
         """
         blocks = make_blocks(self.start_height)
-        with pg.connect(temp_db) as conn:
-            bootstrap_db(conn, blocks)
+        with pg.connect(temp_db_rev1) as conn:
+            # Fill genesis and 1st block only, so we can check if any new blocks got included
+            fill_rev1_db(conn, blocks[0:1])
             yield conn
 
     def test_future_db(self, db: pg.Connection, temp_cfg):
@@ -334,9 +275,9 @@ class TestMigrations:
         with db.cursor() as cur:
             # Check state before run
             cur.execute("select count(*) from core.headers;")
-            assert cur.fetchone()[0] == 1
+            assert cur.fetchone()[0] == 2
             # Fake migration
-            cur.execute("update ew.revision set version = version + 1;")
+            cur.execute("update ew.revision set version = version + 1000;")
         db.commit()
 
         cp = run_watcher(temp_cfg)
@@ -351,26 +292,33 @@ class TestMigrations:
         assert cp.returncode != 0
         with db.cursor() as cur:
             cur.execute("select count(*) from core.headers;")
-            assert cur.fetchone()[0] == 1
+            assert cur.fetchone()[0] == 2
 
     def test_lagging_db_without_m_option(self, db: pg.Connection, temp_cfg):
         """
         Check watcher stops if DB is behind but migrations are not allowed.
         """
+        # Get revision of latest schema
+        with open(SCHEMA_PATH) as f:
+            current_revision = re.findall(
+                r"insert into ew\.revision \(version\) values \((\d+)\)", f.read()
+            )[0]
+            f.read()
+            current_revision = int(current_revision)
+
         with db.cursor() as cur:
             # Check state before run
             cur.execute("select count(*) from core.headers;")
-            assert cur.fetchone()[0] == 1
-
-            # Fake migration
-            cur.execute("update ew.revision set version = version - 1;")
-        db.commit()
+            assert cur.fetchone()[0] == 2
+            cur.execute("select version from ew.revision;")
+            db_revision = cur.fetchone()[0]
+            assert db_revision < current_revision
 
         cp = run_watcher(temp_cfg)
 
         # Check logs
         assert (
-            "Database is 1 revision(s) behind. Run with the -m option to allow migrations to be applied."
+            f"Database is {current_revision - db_revision} revision(s) behind. Run with the -m option to allow migrations to be applied."
             in cp.stdout.decode()
         )
 
@@ -378,28 +326,25 @@ class TestMigrations:
         assert cp.returncode != 0
         with db.cursor() as cur:
             cur.execute("select count(*) from core.headers;")
-            assert cur.fetchone()[0] == 1
+            assert cur.fetchone()[0] == 2
 
-    @pytest.mark.skip("No migrations to test yet")
-    def test_migration_are_applied_if_allowed(self, db: pg.Connection, temp_cfg):
+    # @pytest.mark.skip("No migrations to test yet")
+    def test_migrations_are_applied_if_allowed(self, db: pg.Connection, temp_cfg):
         """
         Check migrations are applied and watcher proceeds normally.
         """
         # Check db is at version 1 initially
         with db.cursor() as cur:
             cur.execute("select version from ew.revision;")
-            assert cur.fetchone[0] == 1
+            assert cur.fetchone()[0] == 1
 
-        cp = run_watcher(temp_cfg)
+        cp = run_watcher(temp_cfg, allow_migrations=True)
 
         # Check logs
-        assert "Applying migration " in cp.stdout.decode()
-
-        # Check db was synced
         assert cp.returncode == 0
+        assert "Applying migration 1 (revision 2)" in cp.stdout.decode()
+
+        # Check migrations are applied
         with db.cursor() as cur:
-            cur.execute("select count(*) from core.headers;")
-            assert cur.fetchone()[0] == 1
-            # And migration was applied
             cur.execute("select version from ew.revision;")
-            assert cur.fetchone()[0] == 2
+            assert cur.fetchone()[0] >= 2