+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Changelog
 
 
-## watcher v0.1.0 - 2022-03-22
-- 🎉 first release
+## v0.2.0 - 2022-04-05
+
+### New features
+[Issue 13](https://github.com/abchrisxyz/ergowatch/issues/13) - Made node poll interval configurable.
+[Issue 16](https://github.com/abchrisxyz/ergowatch/issues/17) - Added token details endpoint.
+[Issue 17](https://github.com/abchrisxyz/ergowatch/issues/17) - Added sync status API.
+[Issue 21](https://github.com/abchrisxyz/ergowatch/issues/21) - Added contracts supply endpoint.
+[Issue 22](https://github.com/abchrisxyz/ergowatch/issues/22) - Added utxo metrics
+
+### Changed
+[Issue 19](https://github.com/abchrisxyz/ergowatch/issues/19) - Token supply endpoint breaks down circulating supply between P2PK and contract addresses.
+
+### Fixed
+[Issue 11](https://github.com/abchrisxyz/ergowatch/issues/11) - Non-zero timestamp for genesis header.
+
+### Housekeeping
+[Issue 12](https://github.com/abchrisxyz/ergowatch/issues/12) - Handle genesis boxes in bootstrapping process (not separately).
+
+
+## v0.1.0 - 2022-03-22
+🎉 first release